refactor(wasm): split leo-wasm into per-command modules mirroring the CLI
Each `commands::*` module corresponds to one `leo` CLI command:

| `leo` command | wasm module          |
|----------------|----------------------|
| `leo build`    | `commands::build`     |
| `leo run`      | `commands::run`       |
| `leo test`     | `commands::test`      |
| `leo fmt`      | `commands::format`    |

Each module exposes the same JSON-returning `*_impl` entry points the
`#[wasm_bindgen]` shim in `wasm_bindings` wraps verbatim — no JS-facing
API change. Shared plumbing (compile boilerplate, manifest parsing,
JSON shaping) moves into a new `wire` module so each command file
focuses on its own command.

Layout:
- `lib.rs` (was 509 LOC, now 119): crate docs (CLI ↔ wasm lookup table)
  + module decls + `wasm_bindings` shim + re-exports.
- `commands/{build,run,test,format}.rs` (~80-200 LOC each).
- `wire.rs` (~160 LOC): `compile_session`/`compile_with`,
  `parse_program_json`/`network_from_*`, `error_json`/`diagnostics_from`/
  `import_summaries`/`clone_file_source`.
- `project.rs` and `evaluate.rs` unchanged.

Adding a new wasm command later (deploy, execute, etc.) is now a
single new file in `commands/` plus one wrapper in `wasm_bindings`.

Verified: `cargo check` native + wasm32, 5/5 leo-wasm tests pass, all
23 Playwright tests pass against the rebuilt wasm.
